+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# KMP Project View Changelog
+# KMP Project Navigator Changelog
 
 ## [1.0.0] - 2026-05-30